requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
ipython==6.4.0
netaddr==0.7.19
paramiko==2.4.1
psycopg2==2.7.4
python-nmap==0.6.1
redis==2.10.6
requests==2.18.4